Don't quit when there are no screens
Default Qt behaviour is to quit when all windows are closed.

Windows can get a close event when the output is being removed,
potentially slightly before we handle the output removed and delete the
window.

This ordering means we can have no screens for enough time for this hook
to take effect.